push para trabajar junto

+++ ubicaciones/src/main/java/backend/grupo130/ubicaciones/dto/deposito/request/DepositoGetByIdRequest.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class DepositoGetByIdRequest {
 
-    private final Integer idDeposito;
+    private final Long idDeposito;
 
 }
