+++ ubicaciones/src/main/java/backend/grupo130/ubicaciones/dto/deposito/response/DepositoGetByIdResponse.java
@@ -9,12 +9,12 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 public class DepositoGetByIdResponse {
 
-    private final Integer idDeposito;
+    private final Long idDeposito;
 
     private final String nombre;
 
     private final BigDecimal costoEstadiaDiario;
 
-    private final Integer idUbicacion;
+    private final Long idUbicacion;
 
 }
